transform sizechart into separate component
